virt: move image utility functions from virt_vm to virt_utils
This allows us to break the circular dependency that currently exists
in virt_vm.py/virt_utils.py.

Reasoning and longer explanation: while testing a script that uses some
features of virt_utils.py, it was noticed that a gstreamer initialization
method was called. That seemed to be unnecessary given that the video
generation code lives in a separate module. A bit more investigation
showed that virt_utils.py imported both virt_vm.py and virt_env_process.py.

This patch removes the circular dependency for virt_vm.py only. The
dependency on virt_env_process.py is a bit more tricky, but will be done.
