+++ client/virt/kvm_vm.py
@@ -741,7 +741,7 @@ class VM(virt_vm.BaseVM):
                 bus, port = get_free_usb_port(image_name, "ehci")
 
             qemu_cmd += add_drive(help,
-                    virt_vm.get_image_filename(image_params, root_dir),
+                    virt_utils.get_image_filename(image_params, root_dir),
                     image_params.get("drive_index"),
                     image_params.get("drive_format"),
                     image_params.get("drive_cache"),
@@ -750,8 +750,8 @@ class VM(virt_vm.BaseVM):
                     image_params.get("drive_serial"),
                     image_params.get("image_snapshot"),
                     image_params.get("image_boot"),
-                    virt_vm.get_image_blkdebug_filename(image_params,
-                                                        self.virt_dir),
+                    virt_utils.get_image_blkdebug_filename(image_params,
+                                                           self.virt_dir),
                     bus,
                     port,
                     image_params.get("bootindex"),
