+++ client/virt/virt_env_process.py
@@ -2,7 +2,7 @@ import os, time, commands, re, logging, glob, threading, shutil
 from autotest_lib.client.bin import utils
 from autotest_lib.client.common_lib import error
 import aexpect, virt_utils, kvm_monitor, ppm_utils, virt_test_setup
-import virt_vm, kvm_vm, libvirt_vm, virt_video_maker
+import virt_vm, kvm_vm, libvirt_vm, virt_video_maker, virt_utils
 try:
     import PIL.Image
 except ImportError:
@@ -24,7 +24,7 @@ def preprocess_image(test, params):
     @param params: A dict containing image preprocessing parameters.
     @note: Currently this function just creates an image if requested.
     """
-    image_filename = virt_vm.get_image_filename(params, test.bindir)
+    image_filename = virt_utils.get_image_filename(params, test.bindir)
 
     create_image = False
 
@@ -35,7 +35,7 @@ def preprocess_image(test, params):
           os.path.exists(image_filename)):
         create_image = True
 
-    if create_image and not virt_vm.create_image(params, test.bindir):
+    if create_image and not virt_utils.create_image(params, test.bindir):
         raise error.TestError("Could not create image")
 
 
@@ -114,13 +114,13 @@ def postprocess_image(test, params):
     """
     if params.get("check_image") == "yes":
         try:
-            virt_vm.check_image(params, test.bindir)
+            virt_utils.check_image(params, test.bindir)
         except Exception, e:
             if params.get("restore_image_on_check_error", "no") == "yes":
-                virt_vm.backup_image(params, test.bindir, 'restore', True)
+                virt_utils.backup_image(params, test.bindir, 'restore', True)
             raise e
     if params.get("remove_image") == "yes":
-        virt_vm.remove_image(params, test.bindir)
+        virt_utils.remove_image(params, test.bindir)
 
 
 def postprocess_vm(test, params, env, name):
