+++ client/virt/virt_utils.py
@@ -10,7 +10,7 @@ import struct, shutil, glob
 from autotest_lib.client.bin import utils, os_dep
 from autotest_lib.client.common_lib import error, logging_config
 from autotest_lib.client.common_lib import logging_manager, git
-from autotest_lib.client.virt import virt_env_process, virt_vm
+from autotest_lib.client.virt import virt_env_process
 from autotest_lib.client.common_lib.syncdata import SyncData, SyncListenServer
 import rss_client, aexpect
 import platform
@@ -539,6 +539,339 @@ def check_kvm_source_dir(source_dir):
         return 2
     else:
         raise error.TestError("Unknown source dir layout, cannot proceed.")
+
+
+# Functions for handling virtual machine image files
+
+def get_image_blkdebug_filename(params, root_dir):
+    """
+    Generate an blkdebug file path from params and root_dir.
+
+    blkdebug files allow error injection in the block subsystem.
+
+    @param params: Dictionary containing the test parameters.
+    @param root_dir: Base directory for relative filenames.
+
+    @note: params should contain:
+           blkdebug -- the name of the debug file.
+    """
+    blkdebug_name = params.get("drive_blkdebug", None)
+    if blkdebug_name is not None:
+        blkdebug_filename = get_path(root_dir, blkdebug_name)
+    else:
+        blkdebug_filename = None
+    return blkdebug_filename
+
+
+def get_image_filename(params, root_dir):
+    """
+    Generate an image path from params and root_dir.
+
+    @param params: Dictionary containing the test parameters.
+    @param root_dir: Base directory for relative filenames.
+
+    @note: params should contain:
+           image_name -- the name of the image file, without extension
+           image_format -- the format of the image (qcow2, raw etc)
+    """
+    image_name = params.get("image_name", "image")
+    image_format = params.get("image_format", "qcow2")
+    if params.get("image_raw_device") == "yes":
+        return image_name
+    if image_format:
+        image_filename = "%s.%s" % (image_name, image_format)
+    else:
+        image_filename = image_name
+    image_filename = get_path(root_dir, image_filename)
+    return image_filename
+
+
+def clone_image(params, vm_name, image_name, root_dir):
+    """
+    Clone master image to vm specific file.
+
+    @param params: Dictionary containing the test parameters.
+    @param vm_name: Vm name.
+    @param image_name: Master image name.
+    @param root_dir: Base directory for relative filenames.
+    """
+    if not params.get("image_name_%s_%s" % (image_name, vm_name)):
+        m_image_name = params.get("image_name", "image")
+        vm_image_name = "%s_%s" % (m_image_name, vm_name)
+        if params.get("clone_master", "yes") == "yes":
+            image_params = params.object_params(image_name)
+            image_params["image_name"] = vm_image_name
+
+            m_image_fn = get_image_filename(params, root_dir)
+            image_fn = get_image_filename(image_params, root_dir)
+
+            logging.info("Clone master image for vms.")
+            utils.run(params.get("image_clone_commnad") % (m_image_fn,
+                                                           image_fn))
+
+        params["image_name_%s_%s" % (image_name, vm_name)] = vm_image_name
+
+
+def rm_image(params, vm_name, image_name, root_dir):
+    """
+    Remove vm specific file.
+
+    @param params: Dictionary containing the test parameters.
+    @param vm_name: Vm name.
+    @param image_name: Master image name.
+    @param root_dir: Base directory for relative filenames.
+    """
+    if params.get("image_name_%s_%s" % (image_name, vm_name)):
+        m_image_name = params.get("image_name", "image")
+        vm_image_name = "%s_%s" % (m_image_name, vm_name)
+        if params.get("clone_master", "yes") == "yes":
+            image_params = params.object_params(image_name)
+            image_params["image_name"] = vm_image_name
+
+            image_fn = get_image_filename(image_params, root_dir)
+
+            logging.debug("Removing vm specific image file %s", image_fn)
+            if os.path.exists(image_fn):
+                utils.run(params.get("image_remove_commnad") % (image_fn))
+            else:
+                logging.debug("Image file %s not found", image_fn)
+
+
+def create_image(params, root_dir):
+    """
+    Create an image using qemu_image.
+
+    @param params: Dictionary containing the test parameters.
+    @param root_dir: Base directory for relative filenames.
+
+    @note: params should contain:
+           image_name -- the name of the image file, without extension
+           image_format -- the format of the image (qcow2, raw etc)
+           image_cluster_size (optional) -- the cluster size for the image
+           image_size -- the requested size of the image (a string
+           qemu-img can understand, such as '10G')
+           create_with_dd -- use dd to create the image (raw format only)
+    """
+    format = params.get("image_format", "qcow2")
+    image_filename = get_image_filename(params, root_dir)
+    size = params.get("image_size", "10G")
+    if params.get("create_with_dd") == "yes" and format == "raw":
+        # maps K,M,G,T => (count, bs)
+        human = {'K': (1, 1),
+                 'M': (1, 1024),
+                 'G': (1024, 1024),
+                 'T': (1024, 1048576),
+                }
+        if human.has_key(size[-1]):
+            block_size = human[size[-1]][1]
+            size = int(size[:-1]) * human[size[-1]][0]
+        qemu_img_cmd = ("dd if=/dev/zero of=%s count=%s bs=%sK"
+                        % (image_filename, size, block_size))
+    else:
+        qemu_img_cmd = get_path(root_dir,
+                                params.get("qemu_img_binary", "qemu-img"))
+        qemu_img_cmd += " create"
+
+        qemu_img_cmd += " -f %s" % format
+
+        image_cluster_size = params.get("image_cluster_size", None)
+        if image_cluster_size is not None:
+            qemu_img_cmd += " -o cluster_size=%s" % image_cluster_size
+
+        qemu_img_cmd += " %s" % image_filename
+
+        qemu_img_cmd += " %s" % size
+
+    utils.system(qemu_img_cmd)
+    return image_filename
+
+
+def remove_image(params, root_dir):
+    """
+    Remove an image file.
+
+    @param params: A dict
+    @param root_dir: Base directory for relative filenames.
+
+    @note: params should contain:
+           image_name -- the name of the image file, without extension
+           image_format -- the format of the image (qcow2, raw etc)
+    """
+    image_filename = get_image_filename(params, root_dir)
+    logging.debug("Removing image file %s", image_filename)
+    if os.path.exists(image_filename):
+        os.unlink(image_filename)
+    else:
+        logging.debug("Image file %s not found", image_filename)
+
+
+def check_image(params, root_dir):
+    """
+    Check an image using the appropriate tools for each virt backend.
+
+    @param params: Dictionary containing the test parameters.
+    @param root_dir: Base directory for relative filenames.
+
+    @note: params should contain:
+           image_name -- the name of the image file, without extension
+           image_format -- the format of the image (qcow2, raw etc)
+
+    @raise VMImageCheckError: In case qemu-img check fails on the image.
+    """
+    vm_type = params.get("vm_type")
+    if vm_type == 'kvm':
+        image_filename = get_image_filename(params, root_dir)
+        logging.debug("Checking image file %s", image_filename)
+        qemu_img_cmd = get_path(root_dir,
+                                params.get("qemu_img_binary", "qemu-img"))
+        image_is_qcow2 = params.get("image_format") == 'qcow2'
+        if os.path.exists(image_filename) and image_is_qcow2:
+            # Verifying if qemu-img supports 'check'
+            q_result = utils.run(qemu_img_cmd, ignore_status=True)
+            q_output = q_result.stdout
+            check_img = True
+            if not "check" in q_output:
+                logging.error("qemu-img does not support 'check', "
+                              "skipping check")
+                check_img = False
+            if not "info" in q_output:
+                logging.error("qemu-img does not support 'info', "
+                              "skipping check")
+                check_img = False
+            if check_img:
+                try:
+                    utils.system("%s info %s" % (qemu_img_cmd, image_filename))
+                except error.CmdError:
+                    logging.error("Error getting info from image %s",
+                                  image_filename)
+
+                cmd_result = utils.run("%s check %s" %
+                                       (qemu_img_cmd, image_filename),
+                                       ignore_status=True)
+                # Error check, large chances of a non-fatal problem.
+                # There are chances that bad data was skipped though
+                if cmd_result.exit_status == 1:
+                    for e_line in cmd_result.stdout.splitlines():
+                        logging.error("[stdout] %s", e_line)
+                    for e_line in cmd_result.stderr.splitlines():
+                        logging.error("[stderr] %s", e_line)
+                    if params.get("backup_image_on_check_error", "no") == "yes":
+                        backup_image(params, root_dir, 'backup', False)
+                    raise error.TestWarn("qemu-img check error. Some bad data "
+                                         "in the image may have gone unnoticed")
+                # Exit status 2 is data corruption for sure, so fail the test
+                elif cmd_result.exit_status == 2:
+                    for e_line in cmd_result.stdout.splitlines():
+                        logging.error("[stdout] %s", e_line)
+                    for e_line in cmd_result.stderr.splitlines():
+                        logging.error("[stderr] %s", e_line)
+                    if params.get("backup_image_on_check_error", "no") == "yes":
+                        backup_image(params, root_dir, 'backup', False)
+                    raise VMImageCheckError(image_filename)
+                # Leaked clusters, they are known to be harmless to data
+                # integrity
+                elif cmd_result.exit_status == 3:
+                    raise error.TestWarn("Leaked clusters were noticed during "
+                                         "image check. No data integrity "
+                                         "problem was found though.")
+
+                # Just handle normal operation
+                if params.get("backup_image", "no") == "yes":
+                    backup_image(params, root_dir, 'backup', True)
+
+        else:
+            if not os.path.exists(image_filename):
+                logging.debug("Image file %s not found, skipping check",
+                              image_filename)
+            elif not image_is_qcow2:
+                logging.debug("Image file %s not qcow2, skipping check",
+                              image_filename)
+
+
+def backup_image(params, root_dir, action, good=True):
+    """
+    Backup or restore a disk image, depending on the action chosen.
+
+    @param params: Dictionary containing the test parameters.
+    @param root_dir: Base directory for relative filenames.
+    @param action: Whether we want to backup or restore the image.
+    @param good: If we are backing up a good image(we want to restore it) or
+            a bad image (we are saving a bad image for posterior analysis).
+
+    @note: params should contain:
+           image_name -- the name of the image file, without extension
+           image_format -- the format of the image (qcow2, raw etc)
+    """
+    def backup_raw_device(src, dst):
+        utils.system("dd if=%s of=%s bs=4k conv=sync" % (src, dst))
+
+    def backup_image_file(src, dst):
+        logging.debug("Copying %s -> %s", src, dst)
+        shutil.copy(src, dst)
+
+    def get_backup_name(filename, backup_dir, good):
+        if not os.path.isdir(backup_dir):
+            os.makedirs(backup_dir)
+        basename = os.path.basename(filename)
+        if good:
+            backup_filename = "%s.backup" % basename
+        else:
+            backup_filename = ("%s.bad.%s" %
+                               (basename, generate_random_string(4)))
+        return os.path.join(backup_dir, backup_filename)
+
+
+    image_filename = get_image_filename(params, root_dir)
+    backup_dir = params.get("backup_dir")
+    if params.get('image_raw_device') == 'yes':
+        iname = "raw_device"
+        iformat = params.get("image_format", "qcow2")
+        ifilename = "%s.%s" % (iname, iformat)
+        ifilename = get_path(root_dir, ifilename)
+        image_filename_backup = get_backup_name(ifilename, backup_dir, good)
+        backup_func = backup_raw_device
+    else:
+        image_filename_backup = get_backup_name(image_filename, backup_dir,
+                                                good)
+        backup_func = backup_image_file
+
+    if action == 'backup':
+        image_dir = os.path.dirname(image_filename)
+        image_dir_disk_free = utils.freespace(image_dir)
+        image_filename_size = os.path.getsize(image_filename)
+        image_filename_backup_size = 0
+        if os.path.isfile(image_filename_backup):
+            image_filename_backup_size = os.path.getsize(image_filename_backup)
+        disk_free = image_dir_disk_free + image_filename_backup_size
+        minimum_disk_free = 1.2 * image_filename_size
+        if disk_free < minimum_disk_free:
+            image_dir_disk_free_gb = float(image_dir_disk_free) / 10**9
+            minimum_disk_free_gb = float(minimum_disk_free) / 10**9
+            logging.error("Dir %s has %.1f GB free, less than the minimum "
+                          "required to store a backup, defined to be 120%% "
+                          "of the backup size, %.1f GB. Skipping backup...",
+                          image_dir, image_dir_disk_free_gb,
+                          minimum_disk_free_gb)
+            return
+        if good:
+            # In case of qemu-img check return 1, we will make 2 backups, one
+            # for investigation and other, to use as a 'pristine' image for
+            # further tests
+            state = 'good'
+        else:
+            state = 'bad'
+        logging.info("Backing up %s image file %s", state, image_filename)
+        src, dst = image_filename, image_filename_backup
+    elif action == 'restore':
+        if not os.path.isfile(image_filename_backup):
+            logging.error('Image backup %s not found, skipping restore...',
+                          image_filename_backup)
+            return
+        logging.info("Restoring image file %s from backup",
+                     image_filename)
+        src, dst = image_filename_backup, image_filename
+
+    backup_func(src, dst)
 
 
 # Functions and classes used for logging into guests and transferring files
@@ -3779,14 +4112,14 @@ def preprocess_images(bindir, params, env):
             vm.destroy(free_mac_addresses=False)
         vm_params = params.object_params(vm_name)
         for image in vm_params.get("master_images_clone").split():
-            virt_vm.clone_image(params, vm_name, image, bindir)
+            clone_image(params, vm_name, image, bindir)
 
 
 def postprocess_images(bindir, params):
     for vm in params.get("vms").split():
         vm_params = params.object_params(vm)
         for image in vm_params.get("master_images_clone").split():
-            virt_vm.rm_image(params, vm, image, bindir)
+            rm_image(params, vm, image, bindir)
 
 
 class MigrationData(object):
